Reformatted code, fixed delay, increased clock speed
Code is being reformatted to the standard stated in github. Increased
the clock speed of the stm32f4 to 180Mhz. This means that the pwm values
will probably be off and need to be readjusted for the motors. Reworked the
delay functions so they are more accurate. Still trying to see if the i2c
is working properly.